Simplify `normalizeErrorResponse`
Over time, we decided that the existing `normalizeErrorResponse`
behavior tried to do way too much.  We've decided that the default
behavior should be to instead just return the error payload that the
server provides, without making modifications to it.

The old behavior can be re-introduced by including the mixin found at
`mixins/legacy/normalize-error-response`.

This also removes the deprecation warnings around the `errors` property
in favor of the replacement `payload` property.

Closes #264
