+++ tests/dummy/app/templates/components/ajax-get.hbs
@@ -1,9 +1,7 @@
-{{#if errors}}
-  <ul class="errors">
-  {{#each errors as |error|}}
-    <li class="error">{{error.title}}</li>
-  {{/each}}
-  </ul>
+{{#if error}}
+  <p class="error">
+    {{error}}
+  </p>
 {{else}}
   {{yield (if isLoaded data (action 'load')) isLoaded}}
 {{/if}}
